+++ src/context/UserContext.tsx
@@ -4,7 +4,7 @@
 import React, { createContext, useState, useEffect, ReactNode } from "react";
 import Cookies from "js-cookie";
 import { User } from "@/types/user";
-import { getUserFromToken } from "@/stores/userService";
+import { getUserFromToken } from "@/services/userService";
 
 interface UserContextType {
   user: User | null;
